requirements.txt
fastapi~=0.111.0
sqlalchemy~=2.0.30
pydantic~=2.5.3
aiofiles~=24.1.0
pydicom~=3.0.1
pillow~=11.2.1
httpx~=0.28.1